Don't duplicate retrieval statuses when generating test data with FactoryGirl.
The :with_work_published_today trait on the :retrieval_status factory will create an associated work which is what we want. 

But, the :work_published_today factory will create its own retrieval_status. Since we are already creating a retrieval_status we can tell the :work_published_today factory to not create any retrieval_statuses. This is only useful when we're not already construction a retrieval status.