Extract CounterByMonthReport from Counter#to_csv.
This further separates the two kinds of reports that Counter#to_csv was generating before. Easier to spec and grok when they're separated.

Still maintaining the Counter#to_csv interface for now.

Note: This does not implement the xml :format count because it is not being captured in MySQL, although it appears "xml_views" was a field captured in CouchDB previously.