Fix failing api/v6/events_controller_spec by sanitizing the work_id/PID using Resolvable#get_id_hash.
When the work_id(s) come thru as query parameters their double slashes in "http://" are preserved fine, but when they come thru as part of the URL path the double slashes get normalized to a single slash.

This makes sure that:
   http:/doi.org/10.1371/journal.pone.000001

which doesn't match anything becomes
   http://doi.org/10.1371/journal.pone.000001

which will make the right work and exhibit the expected behavior.